fix: upgraded tokens not being used
In the previous implementation, the `SessionManagerImpl` was not properly utilizing updated tokens causing session issues. The code has been updated to not retain the first session instance but construct a new one each time, hence always using the newest token. Tests have been added to verify the proper functioning of this feature.