fix: update all users when getting their data [AR-5057] (#2200)
This commit removes distinction between federated and non-federated users in UserRepository when getting known users. All users are now treated in the same manner when getting them from the database. This implies that their data will be refreshed from the API regardless of the user type. Adjustments have been made to the UserRepositoryTest to reflect this change in behavior.