refactor: improve clarity in getting connection requests (#2195)
This commit consists of multiple changes that contribute to clarifying the handling of connections within various components in our architecture.
First, the return type of the `fromDaoToConversationDetails` and `observeConnectionRequestList` functions in the `ConnectionMapper` and `ConnectionRepository` classes respectively, have been updated to explicitly return the `Connection` nested class from `ConversationDetails`. This enhances code readability by better illustrating what specific type of conversation details the function is dealing with.
Second, in the `ConversationScope` class, a deprecated message has been added to the `observeConnectionList` interface, explaining that the name is misleading and not descriptive of its function. This method was replaced by `observePendingConnectionRequestsUseCase`, which better illustrates its function of observing connection requests.
Lastly, `ObserveConnectionListUseCase.kt` was renamed as `ObservePendingConnectionRequestsUseCase.kt`, further enhancing the consistency and readability of our codebase. With these changes, any user can understand that this class is specifically related to observing pending connection requests.

Co-authored-by: Yamil Medina <yamilmedina@users.noreply.github.com>