feat(mls): reevaluate protocol on client demand
Introduced a new UseCase, NotifyConversationIsOpenUseCase, to notify when a conversation is open. This new use case allows the performing of update operations like protocol reevaluation in a lazy manner when a conversation is opened. This addition improves the app's responsiveness when user navigates to it, in case there was no protocol migration before.