refactor(MLS): allow returning values when generating commits. (#2609)
This should cause no changes in behaviour at all, but will allow customisation of some MLS pipelines in the near future.
This refactors the `retryOnCommitFailure` function (renamed to `produceAndSendCommitWithRetryAndResult`), so it allows returning a produced value if the commit succeeds.
The function now takes the responsibility of sending the generated commit, instead of relying on the callers to do so.
It also conveniently provides a MLSClient for the callers, as this is the way to obtain a `CommitBundle`.

(cherry picked from commit 390b880d366f12c84b851dbca2e39fdc1e1c6b6d)