fix(proteus): prevent missing messages by using transactions [WPB-10873] (#2992)
* fix(proteus): prevent missing messages by using transactions

Some clients reported missing messages.
We found that it is not that hard to kill the app by swiping it away between decryption and DB insertion.
CoreCrypto doesn't support transactions yet. So we're only tackling CryptoBox at the moment, but the API changes are adapting CoreCrypto for the future as well.

By not saving the session before inserting the messages into the DB, we can try to process this event again and recover this message.

* test: disable iOS and JS as they don't have transaction support

JS Cryptobox doesn't have it.
iOS uses CoreCrypto

* test: disable JVM

It seems that CryptoBox has static data across instances on JVM, so it can't be tested there either