feat(channel): add new Channel conversation type (#3321)
Transformed `Conversation.Type` from enum to a sealed type, so we can branch groups into Regular and Channel. The same treatment for `ConversationDetails`: Group is split into Regular and Channel.
Adapted the checks from `== Group` to `is Group`, as all currently existing logic for groups should be the same for Channels.

This does not address persistence, nor API support for channels yet.