fix: recover from MLS invalid leaf (#3655)
We were not checking the MLS error messages properly. We need to investigate this with more time and make sure we have the proper error management for MLS, especially when navigating through FFI layer and back, or when it involves MlsTransport errors.

Also, there was a deadlock when starting a MLS Transaction for MLSReset within another MLS Transaction. I am suggesting to make the transaction context mandatory to avoid such things. If used by the App, then another UseCase should be created especifically for the app.